UDT: packet delivery bug fixed + minor changes

+++ 4. SSH Demo/net/udt/include/udt_api.h
@@ -7,11 +7,11 @@
 
 int udt_startup();
 
-int udt_bind   (int socket_fd, const struct sockaddr *addr, socklen_t  len);
-int udt_connect(int socket_fd, const struct sockaddr *addr, socklen_t  len);
+int udt_bind   (int socket_fd, const struct sockaddr *addr, socklen_t len);
+int udt_connect(int socket_fd, const struct sockaddr *addr, socklen_t len);
 
-ssize_t udt_recv(int socket_fd, char *buffer, int len);
-ssize_t udt_send(int socket_fd, char *buffer, int len);
+ssize_t udt_recv(int socket_fd, char *buffer, size_t len);
+ssize_t udt_send(int socket_fd, char *buffer, size_t len);
 
 int udt_close(int socket_fd);
